template/fastapi/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
sqlalchemy
pydantic
pyodbc
Pillow
openai
pytesseract
python-multipart
langchain
langchain_community
flask
newsapi-python
flask-cors
mistralai
serpapi
python-docx
PyPDF2
pdfminer.six
openpyxl
beautifulsoup4
langchain_chroma
langchain_openai
chromadb
donut-python
huggingface_hub
transformers
